fix(scan_filter): declare lower_angle_/upper_angle_ as double
The `lower_angle`/`upper_angle` parameters are declared with a
`double` default (`rclcpp::ParameterValue(0.0)`), but were read into
`float` members. `rclcpp::Parameter::get<float>()` on a parameter
stored as `double` can throw `InvalidParameterTypeException`
depending on the rclcpp version, or silently truncate precision.
Match the member type to the parameter type.

+++ src/scan_filter.cpp
@@ -116,7 +116,7 @@ private:
   rclcpp::Subscription<sensor_msgs::msg::LaserScan>::SharedPtr laser_scan_sub_;
   rclcpp::Publisher<sensor_msgs::msg::LaserScan>::SharedPtr laser_scan_filtered_pub_;
 
-  float lower_angle_, upper_angle_;
+  double lower_angle_, upper_angle_;
 };
 
 /* Main */
